+++ crates/matrix-sdk-common/src/lib.rs
@@ -19,8 +19,6 @@ use std::pin::Pin;
 
 use futures_core::Future;
 #[doc(no_inline)]
-pub use instant;
-#[doc(no_inline)]
 pub use ruma;
 
 pub mod debug;